video stuff windows build.
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@4978 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/include/switch_rtp.h
@@ -348,7 +348,7 @@ SWITCH_DECLARE(int) switch_rtp_write_frame(switch_rtp_t *rtp_session, switch_fra
 */
 SWITCH_DECLARE(int) switch_rtp_write_manual(switch_rtp_t *rtp_session,
 											void *data,
-											uint16_t datalen,
+											uint32_t datalen,
 											uint8_t m, switch_payload_t payload, uint32_t ts, uint16_t mseq, uint32_t ssrc, switch_frame_flag_t *flags);
 
 /*!
